feat: Added support for using core.hooksPath
This adds support to the install.sh script to install the hook templates
to a path which is then configured as core.hooksPath.
The difference is that git will go look in that path for all
repositories (unless a local override of core.hooksPath exists), instead
of copying the templates into each new or cloned repostiry.